minor: moved private multiproc/thread methods to _multioperator
